requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
psycopg2-binary